Move ssl->version to ssl->s3->version
The version is not user configuration but a property of the connection
that we learn partway during the handshake. That state needs to be
reset on SSL_clear, so it's cleaner to implement this by just making it
naturally reset as part of getting a new ssl->s3. (That really should
just be called SSLConnection or something and be a common interface
between TLS and DTLS, but we'll get to that one later.)

Change-Id: I6eaa4ab1df5cbc6569732ed2aa01d84900410080
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71530
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>
(cherry picked from commit 83824d298cf2272a265c44336a493c76d843b93a)
Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@surgut.co.uk>